refactor: Move the setFontSize function out of populateFields.ts into a new src/utils/setFontSize.ts
