+++ src/utils/populateFields.ts
@@ -1,26 +1,10 @@
-import type { PDFForm, PDFFont, PDFTextField } from "pdf-lib";
+import type { PDFForm, PDFFont } from "pdf-lib";
+import { setFontSize } from "./setFontSize";
 import { TextAlignment } from "pdf-lib";
 
 export interface Fonts {
 	arialBold: PDFFont;
 	arialNarrowBold: PDFFont;
-}
-
-function setFontSize(
-	field: PDFTextField,
-	value: string,
-	font: PDFFont,
-	maxSize = 12,
-) {
-	const widgets = field.acroField.getWidgets();
-	const widget = widgets[0];
-	const fieldWidth = widget.getRectangle().width;
-
-	if (font.widthOfTextAtSize(value, maxSize) > fieldWidth) {
-		field.setFontSize(0);
-	} else {
-		field.setFontSize(maxSize);
-	}
 }
 
 export async function populateFields(
